requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    make up
#
attrs==19.3.0             # via -r requirements.in, cattrs
babel==2.8.0              # via flask-babel
cattrs==1.0.0             # via -r requirements.in
certifi==2020.4.5.1       # via requests
chardet==3.0.4            # via requests
click==7.1.2              # via flask
dnspython==1.16.0         # via email-validator
email-validator==1.1.0    # via -r requirements.in
flask-babel==1.0.0        # via -r requirements.in
flask-wtf==0.14.3         # via -r requirements.in
flask==1.1.2              # via -r requirements.in, flask-babel, flask-wtf
gunicorn==20.0.4          # via -r requirements.in
idna==2.9                 # via email-validator, requests
itsdangerous==1.1.0       # via flask, flask-wtf
jinja2==2.11.2            # via flask, flask-babel
libgravatar==0.2.3        # via -r requirements.in
markdown==3.2.2           # via -r requirements.in
markupsafe==1.1.1         # via -r requirements.in, jinja2, wtforms
pyaml==20.4.0             # via -r requirements.in
pyjwt==1.7.1              # via -r requirements.in
python-dotenv==0.13.0     # via -r requirements.in
pytz==2020.1              # via babel, flask-babel
pyyaml==5.3.1             # via pyaml
redis==3.5.2              # via -r requirements.in
requests==2.23.0          # via -r requirements.in
urllib3==1.25.9           # via requests
werkzeug==1.0.1           # via flask
wtforms==2.3.1            # via flask-wtf

# The following packages are considered to be unsafe in a requirements file:
# setuptools
